Updated Plugin Inject
Inject update to allow return statements.

EXTRA NOTE: IMPORTANT: a new variable has been added called callResult to any injected method, if "loadBefore" is set to true (and overwrite is set to false) the aliassed method's return result into  it's "this" by the name callResult, make sure to mention this in the plugin making process on the website!